autogenerate output tests from unit tests

+++ conftest.py
@@ -1,9 +1,29 @@
 """ conftest.py """
 import sys
+import warnings
 from pathlib import Path
+from typing import Iterator
 
 import pytest
 from _pytest.config.argparsing import Parser
+
+
+@pytest.fixture(autouse=True)
+def verify_capsys(
+    capsys: pytest.CaptureFixture[str], request: pytest.FixtureRequest
+) -> Iterator[None]:
+    yield
+    test_name = request.node.name.replace("test_", "")
+    if test_name == "lstm" and sys.version_info < (3, 7):
+        try:
+            verify_output(capsys, f"tests/test_output/{test_name}.out")
+        except AssertionError:
+            warnings.warn(
+                "Verbose output is not determininstic because dictionaries "
+                "are not necessarily ordered in versions before Python 3.7."
+            )
+    else:
+        verify_output(capsys, f"tests/test_output/{test_name}.out")
 
 
 def pytest_addoption(parser: Parser) -> None:
@@ -18,8 +38,10 @@ def verify_output(capsys: pytest.CaptureFixture[str], filename: str) -> None:
     new test_output file.
     """
     captured, _ = capsys.readouterr()
+    filepath = Path(filename)
+    if not captured and not filepath.exists():
+        return
     if "--overwrite" in sys.argv:
-        filepath = Path(filename)
         filepath.parent.mkdir(exist_ok=True)
         filepath.touch(exist_ok=True)
         filepath.write_text(captured)
